requirements.txt
flask>=3.0.0
MetaTrader5>=5.0.45
pandas>=2.0.0
pandas_ta>=0.3.14b
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
